calc dix

+++ betamodel.py
@@ -393,7 +393,7 @@ def get_house_price_data(consumption_percentage, savings, age, income):
 
     SOaccumulated_wealth_at_67 = int(round(SOaccumulated_wealth_at_67/1000))*1000
     accumulated_wealth_at_67 = int(round(accumulated_wealth_at_67/1000))*1000
-    transformed_wealth_data = (round(transformed_wealth_data/1000))*1000
+    transformed_wealth_data = int(round(transformed_wealth_data/1000))*1000
 
     results = {
         "affordability_status": df['Affordability Status'].iloc[-1],
